Added minimap to scene03 Edited abit on scene02 Organised abit in image folder

+++ DM2122_Framework/Application/Source/SceneMenu.h
@@ -82,7 +82,7 @@ private:
 	void RenderMesh(Mesh *mesh, bool enableLight);
 	void RenderText(Mesh* mesh, std::string text, Color color);
 	void RenderTextOnScreen(Mesh* mesh, std::string text, Color color, float size, float x, float y);
-	void RenderMeshOnScreen(Mesh* mesh, int x, int y, int sizex, int sizey);
+	void RenderMeshOnScreen(Mesh* mesh, int x, int y, float sizex, float sizey);
 
 	bool enableLight = true;
